face_auth_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
pymongo==4.6.1
python-dotenv==1.0.0
numpy==1.26.3 